Move system.runtime's tests to system.runtime package
