+++ presto-tests/src/test/java/io/prestosql/connector/system/runtime/TestKillQuery.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.connector.system;
+package io.prestosql.connector.system.runtime;
 
 import io.prestosql.Session;
 import io.prestosql.plugin.tpch.TpchPlugin;
